admintest: move admin test into new package
