+++ executor/admintest/BUILD.bazel
@@ -1,0 +1,38 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "admintest_test",
+    timeout = "short",
+    srcs = [
+        "admin_test.go",
+        "main_test.go",
+    ],
+    flaky = True,
+    shard_count = 23,
+    deps = [
+        "//config",
+        "//domain",
+        "//errno",
+        "//kv",
+        "//meta/autoid",
+        "//parser/model",
+        "//planner/core",
+        "//session",
+        "//sessionctx/stmtctx",
+        "//sessionctx/variable",
+        "//table",
+        "//table/tables",
+        "//testkit",
+        "//testkit/testsetup",
+        "//testkit/testutil",
+        "//types",
+        "//util/codec",
+        "//util/logutil/consistency",
+        "//util/mock",
+        "@com_github_pingcap_errors//:errors",
+        "@com_github_stretchr_testify//require",
+        "@com_github_tikv_client_go_v2//tikv",
+        "@org_uber_go_goleak//:goleak",
+        "@org_uber_go_zap//:zap",
+    ],
+)
